+++ R/get_c14data.R
@@ -144,6 +144,7 @@ get_all_parser_functions <- function() {
     "14cpalaeolithic" = c14bazAAR::get_14cpalaeolithic,
     "medafricarbon" = c14bazAAR::get_medafricarbon,
     "jomon" = c14bazAAR::get_jomon,
+    "mesorad" = c14bazAAR::get_mesorad,
     "katsianis" = c14bazAAR::get_katsianis
   ))
 }