+++ R/get_eubar.R
@@ -3,46 +3,44 @@
 get_eubar <- function(db_url = get_db_url("eubar")) {
 
   check_if_packages_are_available("readxl")
-
   check_connection_to_url(db_url)
 
-  # download data to temporary file
-  tempo <- tempfile()
-  utils::download.file(db_url, tempo, mode = "wb", quiet = TRUE)
+  # download data
+  temp <- tempfile(fileext = ".xlsx")
+  utils::download.file(db_url, temp, mode = "wb", quiet = TRUE)
 
   # read data
-  eubar <- tempo %>%
+  db_raw <- temp %>%
     readxl::read_excel(
-      na = c("Combination fails", "nd", "-"),
-    ) %>%
-    dplyr::mutate_if(
-      sapply(., is.character),
-      trimws
-    ) %>%
+      sheet = 1,
+      col_types = "text",
+      na = c("-"),
+      trim_ws = TRUE
+    )
+
+  # delete temporary file
+  unlink(temp)
+
+  # final data preparation
+  eubar <- db_raw %>%
     dplyr::transmute(
       labnr = .[[13]],
       c14age = .[[14]],
       c14std = .[[15]],
-      c13val = NA,
       material = .[[18]],
       country = .[[10]],
       region = .[[9]],
       site = .[[1]],
       lat = .[[12]],
       lon = .[[11]],
-      period = NA,
       feature = .[[19]],
-      shortref = .[[22]],
-      comment = NA
+      shortref = .[[22]]
     ) %>%
     dplyr::mutate(
       sourcedb = "eubar",
       sourcedb_version = get_db_version("eubar")
     ) %>%
     as.c14_date_list()
-
-  # delete temporary file
-  unlink(tempo)
 
   return(eubar)
 }